Move TransferProgress to its own file
